Add `useLeafModelState` hook
